pyproject.toml
[project]
name = "saas-sec-agents"
version = "0.2.0"
description = "SaaS Security multi-agent AI system for OSCAL/SSCF assessments"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0",
  "simple-salesforce>=1.12.6",
  "click>=8.1.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.0",
  "docxtpl>=0.18.0",
  "defusedxml>=0.7.1",
  "mem0ai>=0.1.0",
  "qdrant-client>=1.10.0",
]

[project.scripts]
sfdc-connect = "skills.sfdc_connect.sfdc_connect:cli"
oscal-assess = "skills.oscal_assess.oscal_assess:cli"
sscf-benchmark = "skills.sscf_benchmark.sscf_benchmark:cli"
agent-loop = "harness.loop:cli"

[tool.setuptools.packages.find]
include = ["skills*", "harness*"]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6.0",
  "pytest>=8.0.0",
  "pytest-mock>=3.14.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
